[#8954] Rename data fetching constants module to util
